+++ integration/kubernetes/helm-chart/alluxio/CHANGELOG.md
@@ -305,3 +305,8 @@
 0.6.50
 
 - Remove dummy example nginx from CSI
+
+0.6.51
+
+- Remove alluxio-fuse-client generation from helm-generate.sh script
+- Fix global mount indentation issue in master